kernel: optionally apply patch

+++ app-kernel/stable-sources/stable-sources-6.11.9999.ebuild
@@ -21,6 +21,8 @@ LICENSE="GPL"
 SLOT="$(ver_cut 1-2)"
 KEYWORDS="amd64 arm64"
 
+IUSE="mediatek"
+
 RESTRICT="network-sandbox"
 
 DEPEND="
@@ -31,13 +33,16 @@ DEPEND="
 src_prepare() {
 	kernel-2_src_prepare
 
-	eapply "${FILESDIR}"/mt7925.patch
 	eapply "${FILESDIR}"/reg.patch
 
-	local mt76_dir="${WORKDIR}/mt76"
-	git clone https://github.com/openwrt/mt76.git "${mt76_dir}" --depth 1 || die "Failed to clone mt76 repository"
-	rm -rf "${mt76_dir}"/.git
-	cp -r "${mt76_dir}" "${S}/drivers/net/wireless/mediatek/" || die "Failed to copy new mt76 directory"
-	eapply "${FILESDIR}"/mt76.patch
+	if use mediatek ; then
+		eapply "${FILESDIR}"/mt7925.patch
+
+		local mt76_dir="${WORKDIR}/mt76"
+		git clone https://github.com/openwrt/mt76.git "${mt76_dir}" --depth 1 || die "Failed to clone mt76 repository"
+		rm -rf "${mt76_dir}"/.git
+		cp -r "${mt76_dir}" "${S}/drivers/net/wireless/mediatek/" || die "Failed to copy new mt76 directory"
+		eapply "${FILESDIR}"/mt76.patch
+	fi
 }
